🗽 Operate over indices, not elements.
Interestingly, this slightly increases heap usage. This is likely
because the readings array is simply an array of pointers to readings
arrays, which themselves stay in place the entire time.

The real play for in-place efficiency would be to operate over reading
strings in place.

(Tried Set too, and it had even more overhead, for the same reason.)